+++ NEWS.md
@@ -1,3 +1,10 @@
+# Changes in v1.0.0
+
+* Add `max_count` to `textmodel_lss.fcm()` that will be passed to `x_max` in `rsparse::GloVe$new()`.
+* Add `max_words` to `textplot_terms()` to avoid overcrowding.
+* Make `textplot_terms()` work with objects from `textmodel_lss.fcm()`.
+* Add `concatenator` to `as.seedwords()`.
+
 # Changes in v0.9.9
 
 * Correct how `textstat_context()` and `char_context()` computes statistics.